DAVIS: abstract common operations, eliminate code duplication between normal and RPi variant. DAVIS port complete, RPi WIP.
